feat: オンラインの背景忘れとキャンバスのクラス調整
